+++ source/features/highlight-deleted-and-added-files-in-diffs.tsx
